+++ src/helperComponents/IconDynamicComponent/constants.ts
@@ -8,5 +8,6 @@ export const IconComponentMapping: { [key: string]: any } = {
   ),
   errorCircleFilled: lazy((): any => import('../../components/SVGIcons/IconErrorCircleFilled')),
   infoFilled: lazy((): any => import('../../components/SVGIcons/IconInfoFilled')),
-  checkmark: lazy((): any => import('../../components/SVGIcons/IconCheckmark'))
+  checkmark: lazy((): any => import('../../components/SVGIcons/IconCheckmark')),
+  subtract: lazy((): any => import('../../components/SVGIcons/IconSubtract'))
 }