+++ src/stories/Select.stories.tsx
@@ -3,17 +3,16 @@ import {
   Button,
   Select as _Select,
   MultiSelect as _MultiSelect,
-  // FilterSelect as _FilterDropdown,
   NestedSelect as _NestedSelect,
   ButtonSelect as _ButtonSelect,
-  Menu
+  ProfileDropdown as _ProfileDropdown,
+  Menu,
+  Avatar
 } from '../index'
 // @ts-ignore
 import imageFile from '../assets/images/avatar.jpg'
 import IconPerson from '../components/SVGIcons/IconPerson'
 import IconInfo from '../components/SVGIcons/IconInfo'
-import IconCaretDownFilled from '../components/SVGIcons/IconCaretDownFilled'
-import IconSelectAllOff from '../components/SVGIcons/IconSelectAllOff'
 import IconMore from '../components/SVGIcons/IconMore'
 import { StoryFn } from '@storybook/react'
 import {
@@ -22,6 +21,9 @@ import {
   TSingleSelectPropTypes
 } from '../components/Select/types'
 import { Positions } from '../components/Tooltip/types'
+import IconSignOut from '../components/SVGIcons/IconSignOut'
+import IconQuestion from '../components/SVGIcons/IconQuestion'
+import IconChatMultiple from '../components/SVGIcons/IconChatMultiple'
 
 type TItemValue = string | number | null
 declare type TItemLabel = string
@@ -451,6 +453,55 @@ NestedSelect.args = {
   ),
   labelAddons: <IconInfo size={'xsmall'} type={'information'} className={'ml-4'} />
 }
+
+// ----------PROFILE DROPDOWN---------------
+
+const ProfileDropdownTemplate = (args: any): JSX.Element => {
+  return (
+    <div
+      style={{
+        width: 320,
+        display: 'flex',
+        flexDirection: 'column',
+        position: 'absolute',
+        left: 300
+      }}
+    >
+      <div style={{ position: 'absolute', left: 300 }}>
+        <_ProfileDropdown {...args} />
+      </div>
+    </div>
+  )
+}
+
+export const ProfileDropdown = ProfileDropdownTemplate.bind({})
+
+// @ts-ignore
+ProfileDropdown.args = {
+  avatar: <Avatar size={'medium'} initials={'RG'} />,
+  name: 'Mariam Yeghiazaryan',
+  email: 'mariam.yeghiazaryan@gmail.com',
+  bodyItems: [
+    {
+      icon: <IconQuestion size={'small'} className={'mr-8'} />,
+      text: 'Հաճախ տրվող հարցեր',
+      url: '#s'
+    },
+    {
+      icon: <IconChatMultiple size={'small'} className={'mr-8'} />,
+      text: 'Հետադարձ կապ',
+      url: '#s'
+    }
+  ],
+  footerItems: [
+    {
+      icon: <IconSignOut size={'small'} className={'mr-8'} />,
+      text: 'Ելք',
+      onclick: () => alert('button clicked')
+    }
+  ]
+}
+
 // ----------FILTERDROPDOWN---------------
 
 // const FilterDropdownTemplate = (args: any): JSX.Element => {